requirements.txt
# requirements.txt — v1.0.0
paramiko==3.4.0
pyyaml==6.0.1
requests==2.31.0